Fixed include errors Created StationData.hpp and .cpp Farmers now produce food Create base Job class Added unassigning workers

+++ src/Resource.hpp
@@ -5,7 +5,7 @@
 class Resource
 {
 public:
-	int amount = 0;
+	float amount = 0;
 	bool unlocked = false;
 
 	virtual std::string getName() = 0;
